REFACTOR: se creo la carpeta procesos en utils y se movieron los scripts de procesos a esa carpeta
