+++ Vista/pages/modulo_procesos.html
@@ -43,7 +43,7 @@
       type="text/css"
       href="/node_modules/datatables.net-bs5/css/dataTables.bootstrap5.min.css"
     />
-    <script defer type="module" src="src/utils/Procesos.js"></script>
+    <script defer type="module" src="src/utils/procesos/Procesos.js"></script>
   </head>
 
   <body>
@@ -365,6 +365,9 @@
       integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
       crossorigin="anonymous"
     ></script>
-    <script type="module" src="src/utils/ModificarProcesos.js"></script>
+    <script
+      type="module"
+      src="src/utils/procesos/ModificarProcesos.js"
+    ></script>
   </body>
 </html>
