+++ Vista/pages/registros/registro_procesos.html
@@ -93,7 +93,7 @@
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
     integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
     crossorigin="anonymous"></script>
-  <script defer type="module" src="../src/utils/RegistroProcesos.js"></script>
+  <script defer type="module" src="../src/utils/registro/RegistroProcesos.js"></script>
   <script src="../src/utils/Env.js"></script>
 </body>
